requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
sqlalchemy>=2.0.35
psycopg2-binary>=2.9.5
# passlib[bcrypt]>=1.7.4 (Replaced by direct bcrypt usage for Py3.14 stability)
PyJWT>=2.9.0
python-dotenv>=1.0.1
pydantic[email]>=2.10.0
httpx>=0.27.2
pytest>=8.3.3
alembic>=1.13.3
python-multipart>=0.0.12
bcrypt>=4.0.1
anyio>=4.0.0
